# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tiza API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/integrations/nema/readiness/request": {
      "post": {
        "tags": [
          "nema"
        ],
        "summary": "Request Readiness",
        "operationId": "request_readiness_api_integrations_nema_readiness_request_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/integrations/nema/readiness": {
      "post": {
        "tags": [
          "nema"
        ],
        "summary": "Receive Readiness",
        "operationId": "receive_readiness_api_integrations_nema_readiness_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssertionAccept"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/integrations/nema/grants/{grant_id}": {
      "delete": {
        "tags": [
          "nema"
        ],
        "summary": "Delete Grant",
        "operationId": "delete_grant_api_integrations_nema_grants__grant_id__delete",
        "parameters": [
          {
            "name": "grant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Grant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/attempts/{attempt_id}/receipt": {
      "get": {
        "tags": [
          "nema"
        ],
        "summary": "Get Receipt",
        "operationId": "get_receipt_api_attempts__attempt_id__receipt_get",
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Attempt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/demo/clock/advance": {
      "post": {
        "summary": "Advance Clock",
        "operationId": "advance_clock_api_demo_clock_advance_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvanceClock"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/cycles/{cycle_id}/cancel-reminders": {
      "post": {
        "summary": "Cancel Reminders",
        "operationId": "cancel_reminders_api_cycles__cycle_id__cancel_reminders_post",
        "parameters": [
          {
            "name": "cycle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cycle Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdvanceClock": {
        "properties": {
          "hours": {
            "type": "integer",
            "maximum": 72.0,
            "minimum": 1.0,
            "title": "Hours"
          }
        },
        "type": "object",
        "required": [
          "hours"
        ],
        "title": "AdvanceClock"
      },
...
      "AssertionAccept": {
        "properties": {
          "token": {
            "type": "string",
            "maxLength": 65536,
            "minLength": 10,
            "title": "Token"
          },
          "confirm_unknown_key": {
            "type": "boolean",
            "title": "Confirm Unknown Key",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "token"
        ],
        "title": "AssertionAccept"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RequestCreate": {
        "properties": {
          "cycle_id": {
            "type": "string",
            "title": "Cycle Id"
          },
          "purpose": {
            "type": "string",
            "maxLength": 200,
            "minLength": 3,
            "title": "Purpose",
            "default": "Plan the next approved practice"
          },
          "requirements": {
            "items": {
              "$ref": "#/components/schemas/Requirement"
            },
            "type": "array",
            "maxItems": 12,
            "minItems": 1,
            "title": "Requirements"
          }
        },
        "type": "object",
        "required": [
          "cycle_id",
          "requirements"
        ],
        "title": "RequestCreate"
      },
      "Requirement": {
        "properties": {
          "concept": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Concept"
          },
          "ability": {
            "type": "string",
            "enum": [
              "recognize",
              "retrieve",
              "explain",
              "apply",
              "transfer",
              "discriminate"
            ],
            "title": "Ability"
          }
        },
        "type": "object",
        "required": [
          "concept",
          "ability"
        ],
        "title": "Requirement"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}